feature: :sparkles: valida permissões

+++ app/Http/Controllers/EstoqueController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Models\Estoque;
+use Illuminate\Container\Attributes\Auth;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth as FacadesAuth;
+use App\Models\User;
 
 class EstoqueController extends Controller
 {
@@ -15,6 +18,13 @@ class EstoqueController extends Controller
 
     public function criarEstoque(Request $request)
     {
+
+        $role = User::find(FacadesAuth::id())->role;
+
+        if($role != 'admin') {
+            return response()->json(['message' => 'Você não tem permissão para criar um item de estoque!'], 403);
+        }
+
         $estoque = new Estoque();
         $estoque->nome = $request->nome;
         $estoque->categoria = $request->categoria;
@@ -26,12 +36,26 @@ class EstoqueController extends Controller
 
     public function deletarEstoque($id)
     {
+
+        $role = User::find(FacadesAuth::id())->role;
+
+        if($role != 'admin') {
+            return response()->json(['message' => 'Você não tem permissão para deletar um item de estoque!'], 403);
+        }
+
         Estoque::find($id)->delete();
         return response()->json(['message' => 'Item deletado com sucesso!']);
     }
 
     public function atualizar(Request $request, $id)
     {
+
+        $role = User::find(FacadesAuth::id())->role;
+
+        if($role != 'admin') {
+            return response()->json(['message' => 'Você não tem permissão para atualizar um item de estoque!'], 403);
+        }
+
         $estoque = Estoque::find($id);
         $estoque->nome = $request->nome;
         $estoque->categoria = $request->categoria;
@@ -41,23 +65,9 @@ class EstoqueController extends Controller
         return response()->json($estoque);
     }
 
-    public function atualizarQuantidadeEstoque(Request $request, $id)
-    {
-        $estoque = Estoque::find($id);
-
-        if ($request->action == 'adicionar') {
-            $estoque->quantidade += 1;
-        } else {
-            $estoque->quantidade -= 1;
-        }
-
-        $estoque->save();
-
-        return response()->json($estoque);
-    }
-
     public function consumirQuantidadeEstoque(Request $request, $id)
     {
+
         $estoque = Estoque::find($id);
 
         if (!$estoque) {
